Type checking (#9)
* UTW

* UTS

* word, dword checks

* Utm

* UTI

* UTE UTC UTD

* Jrl

* Itp

* Ifo

* are

* Dlg

* Fac

* fac, gic

* cargo format

* git

* git.rs merge

---------

+++ src/fac.rs
@@ -26,7 +26,7 @@ pub struct Faction {
 
   /// Field representing the faction global value.
   #[serde(skip_serializing_if = "Option::is_none", rename = "FactionGlobal")]
-  pub faction_global: Option<NwValue<u8>>,
+  pub faction_global: Option<NwValue<u16>>,
 
   /// Field representing the faction name.
   #[serde(skip_serializing_if = "Option::is_none", rename = "FactionName")]
@@ -46,13 +46,13 @@ pub struct Rep {
 
   /// Field representing the first faction ID.
   #[serde(skip_serializing_if = "Option::is_none", rename = "FactionID1")]
-  pub faction_id1: Option<NwValue<u8>>,
+  pub faction_id1: Option<NwValue<u32>>,
 
   /// Field representing the second faction ID.
   #[serde(skip_serializing_if = "Option::is_none", rename = "FactionID2")]
-  pub faction_id2: Option<NwValue<u8>>,
+  pub faction_id2: Option<NwValue<u32>>,
 
   /// Field representing the faction reputation value.
   #[serde(skip_serializing_if = "Option::is_none", rename = "FactionRep")]
-  pub faction_rep: Option<NwValue<u8>>,
+  pub faction_rep: Option<NwValue<u32>>,
 }
